+++ src/game/client/components/menus.cpp
@@ -1004,8 +1004,8 @@ bool CMenus::OnMouseMove(float x, float y)
 	m_MousePos.y += y;
 	if(m_MousePos.x < 0) m_MousePos.x = 0;
 	if(m_MousePos.y < 0) m_MousePos.y = 0;
-	if(m_MousePos.x > Graphics()->ScreenWidth()) m_MousePos.x = Graphics()->ScreenWidth();
-	if(m_MousePos.y > Graphics()->ScreenHeight()) m_MousePos.y = Graphics()->ScreenHeight();
+	if(m_MousePos.x > Graphics()->ScreenWidth()) m_MousePos.x = (float)Graphics()->ScreenWidth();
+	if(m_MousePos.y > Graphics()->ScreenHeight()) m_MousePos.y = (float)Graphics()->ScreenHeight();
 	
 	return true;
 }
